new code to switch wrapping and upgrading
