+++ package.json
@@ -20,7 +20,8 @@
     "next": "^14.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "tailwindcss-animate": "^1.0.7"
+    "tailwindcss-animate": "^1.0.7",
+    "viem": "^2.16.2"
   },
   "engines": {
     "node": ">=18.17.0"
